Add files via upload
I followed all instructions correctly and double checked my work but my "All Stocks Analysis" still does not look like the screenshot provided. I believe the error is somewhere in '6)Output data for current ticker but it looks exactly the same as the code provided in 2.3.3. Why is it still giving me an error code when the code is appearing correct?